fix(market-server-rust): unsufficient wallet cash numerical precision
The cash type of the wallet ORM entity was defined to be a numerical with a precision of 5 for a scale of 2 which was unsufficient for the initial cash at bots enrollment ($1000). The numerical precision was increased to 15 to avoid any other error.

+++ servers/rust-server/entity/wallet.rs
@@ -8,10 +8,9 @@ pub struct Model {
     #[sea_orm(primary_key)]
     pub id: i32,
     pub bot_id: i32,
-    #[sea_orm(column_type = "Decimal(Some((5, 2)))")]
+    #[sea_orm(column_type = "Decimal(Some((15, 2)))")]
     pub cash: Decimal,
-    #[sea_orm(column_type = "Decimal(Some((5, 2)))")]
-    pub asset: Decimal,
+    pub asset: i32,
     pub created_at: DateTimeWithTimeZone,
     pub updated_at: DateTimeWithTimeZone,
 }
